Move TableTypes to a new package
